fixes from meeting
